@@ -14,12 +14,15 @@
     "author": "Roman Dvornov <rdvornov@gmail.com> (https://github.com/lahmatiy)",
     "license": "MIT",
     "repository": "discoveryjs/json-ext",
-    "main": "./src/index",
+    "engines": {
+        "node": ">=14.17.0"
+    },
+    "type": "module",
+    "main": "./src/index.js",
     "browser": {
         "./src/stringify-info.js": "./src/stringify-info.js",
         "./src/parse-chunked.js": "./src/parse-chunked.js",
         "./src/stringify-stream.js": "./src/stringify-stream-browser.js",
-        "./src/text-decoder.js": "./src/text-decoder-browser.js",
         "./src/version.js": "./dist/version.js"
     },
     "types": "./index.d.ts",
@@ -46,9 +49,6 @@
         "mocha": "^9.2.1",
         "rollup": "^2.28.2",
         "rollup-plugin-terser": "^7.0.2"
-    },
-    "engines": {
-        "node": ">=10.0.0"
     },
     "files": [
         "dist",
