@@ -1,3 +1,3 @@
-module.exports = () => {
+export function stringifyStream() {
     throw new Error('Method is not supported');
 };
