@@ -1,6 +1,4 @@
-module.exports = {
-    version: require('./version'),
-    stringifyInfo: require('./stringify-info'),
-    stringifyStream: require('./stringify-stream'),
-    parseChunked: require('./parse-chunked')
-};
+export { version } from './version.js';
+export { stringifyInfo } from './stringify-info.js';
+export { stringifyStream } from './stringify-stream.js';
+export { parseChunked } from './parse-chunked.js';
